rename subpackages of mp4tool
