+++ mp4tool/divide/divide.go
@@ -1,4 +1,4 @@
-package mp4divide
+package divide
 
 import (
 	"fmt"
